refactor(scanner): Modularize tests more
Extract several classes as separate test fixtures for a better overview
and to allow code sharing.

As part of that,

- get rid of the "org.ossreviewtoolkit.scanner.scanners" package in tests
  as no such package exists in the "main" source set,
- combine two `createScanner()` functions (which requires a minor change
  to some expected results),
- move `MultipleScannersTest` to the "test" source set as it now really is
  a non-functional test in the sense that it does not depend on a VCS
  tool for provenance resolution anymore.

Note that the "test" source set is automatically configured to depend on
the "testFixtures" source set of the same module [1], so only the custom
"funTest" source set needs to have the dependency added.

[1]: https://github.com/gradle/gradle/blob/v9.1.0/platforms/jvm/plugins-jvm-test-fixtures/src/main/java/org/gradle/api/plugins/JavaTestFixturesPlugin.java#L92-L95
