+++ scanner/src/funTest/kotlin/provenance/DefaultNestedProvenanceResolverFunTest.kt
@@ -221,10 +221,3 @@ class DefaultNestedProvenanceResolverFunTest : WordSpec() {
         }
     }
 }
-
-internal class DummyNestedProvenanceStorage : NestedProvenanceStorage {
-    override fun readNestedProvenance(root: RepositoryProvenance): NestedProvenanceResolutionResult? = null
-    override fun writeNestedProvenance(root: RepositoryProvenance, result: NestedProvenanceResolutionResult) {
-        /** no-op */
-    }
-}
