+++ scanner/src/funTest/resources/scanner-integration-all-pkgs-expected-ort-result.yml
@@ -197,6 +197,8 @@ scanner:
     skip_concluded: false
     skip_excluded: false
     include_files_without_findings: false
+    archive:
+      enabled: false
     detected_license_mapping:
       LicenseRef-scancode-agpl-generic-additional-terms: "NOASSERTION"
       LicenseRef-scancode-free-unknown: "NOASSERTION"
