+++ scanner/src/testFixtures/kotlin/provenance/DummyProvenanceStorage.kt
@@ -1,0 +1,45 @@
+/*
+ * Copyright (C) 2021 The ORT Project Authors (see <https://github.com/oss-review-toolkit/ort/blob/main/NOTICE>)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+package org.ossreviewtoolkit.scanner.provenance
+
+import org.ossreviewtoolkit.model.Identifier
+import org.ossreviewtoolkit.model.RemoteArtifact
+import org.ossreviewtoolkit.model.VcsInfo
+
+class DummyProvenanceStorage : PackageProvenanceStorage {
+    override fun readProvenance(id: Identifier, sourceArtifact: RemoteArtifact): PackageProvenanceResolutionResult? =
+        null
+
+    override fun readProvenance(id: Identifier, vcs: VcsInfo): PackageProvenanceResolutionResult? = null
+
+    override fun readProvenances(id: Identifier): List<PackageProvenanceResolutionResult> = emptyList()
+
+    override fun writeProvenance(id: Identifier, vcs: VcsInfo, result: PackageProvenanceResolutionResult) {
+        /* no-op */
+    }
+
+    override fun writeProvenance(
+        id: Identifier,
+        sourceArtifact: RemoteArtifact,
+        result: PackageProvenanceResolutionResult
+    ) { /* no-op */ }
+
+    override fun deleteProvenances(id: Identifier) { /* no-op */ }
+}
