+++ scanner/src/testFixtures/kotlin/provenance/FakeProvenanceDownloader.kt
@@ -1,0 +1,37 @@
+/*
+ * Copyright (C) 2021 The ORT Project Authors (see <https://github.com/oss-review-toolkit/ort/blob/main/NOTICE>)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+package org.ossreviewtoolkit.scanner.provenance
+
+import java.io.File
+
+import org.ossreviewtoolkit.model.KnownProvenance
+import org.ossreviewtoolkit.model.toYaml
+import org.ossreviewtoolkit.utils.ort.createOrtTempDir
+
+/**
+ * An implementation of [ProvenanceDownloader] that creates a file called [filename] containing the serialized
+ * provenance, instead of actually downloading the source code.
+ */
+class FakeProvenanceDownloader(val filename: String = "fake.txt") : ProvenanceDownloader {
+    override fun download(provenance: KnownProvenance): File =
+        createOrtTempDir().apply {
+            resolve(filename).writeText(provenance.toYaml())
+        }
+}
